Move toEntityKey/toDtoKey mappers into Mapper files
