+++ feat/findings/be/driven/impl/src/main/kotlin/cz/adamec/timotej/snag/findings/be/driven/impl/internal/FindingTypeEntityKey.kt
@@ -12,17 +12,8 @@
 
 package cz.adamec.timotej.snag.findings.be.driven.impl.internal
 
-import cz.adamec.timotej.snag.feat.findings.business.FindingType
-
 internal enum class FindingTypeEntityKey {
     CLASSIC,
     UNVISITED,
     NOTE,
 }
-
-internal fun FindingType.toEntityKey(): FindingTypeEntityKey =
-    when (this) {
-        is FindingType.Classic -> FindingTypeEntityKey.CLASSIC
-        is FindingType.Unvisited -> FindingTypeEntityKey.UNVISITED
-        is FindingType.Note -> FindingTypeEntityKey.NOTE
-    }
