+++ feat/findings/be/driving/impl/src/main/kotlin/cz/adamec/timotej/snag/findings/be/driving/impl/internal/FindingTypeDtoKey.kt
@@ -12,17 +12,8 @@
 
 package cz.adamec.timotej.snag.findings.be.driving.impl.internal
 
-import cz.adamec.timotej.snag.feat.findings.business.FindingType
-
 internal enum class FindingTypeDtoKey {
     CLASSIC,
     UNVISITED,
     NOTE,
 }
-
-internal fun FindingType.toDtoKey(): FindingTypeDtoKey =
-    when (this) {
-        is FindingType.Classic -> FindingTypeDtoKey.CLASSIC
-        is FindingType.Unvisited -> FindingTypeDtoKey.UNVISITED
-        is FindingType.Note -> FindingTypeDtoKey.NOTE
-    }
