+++ feat/findings/fe/driven/impl/src/commonMain/kotlin/cz/adamec/timotej/snag/findings/fe/driven/internal/db/FindingTypeEntityKey.kt
@@ -12,17 +12,8 @@
 
 package cz.adamec.timotej.snag.findings.fe.driven.internal.db
 
-import cz.adamec.timotej.snag.feat.findings.business.FindingType
-
 internal enum class FindingTypeEntityKey {
     CLASSIC,
     UNVISITED,
     NOTE,
 }
-
-internal fun FindingType.toEntityKey(): FindingTypeEntityKey =
-    when (this) {
-        is FindingType.Classic -> FindingTypeEntityKey.CLASSIC
-        is FindingType.Unvisited -> FindingTypeEntityKey.UNVISITED
-        is FindingType.Note -> FindingTypeEntityKey.NOTE
-    }
